recipe_utils , tasks, utils changed

+++ derivative-recipe-generatorq/tasks/tasks.py
@@ -34,7 +34,7 @@ search_url = "{0}?query={{\"filter\": {{\"bag\": \"{1}\"}}}}"
 bagList=[]
 
 def getAllBags():
-    response = requests.get('https://cc.lib.ou.edu/api/catalog/data/catalog/digital_objects/?query={"filter":{"department":"DigiLab","project":{"$ne":"private"},"locations.s3.exists":{"$eq":true},"derivatives.jpeg_040_antialias.recipe":{"$exists":false,"$error":ne}}}&format=json&page_size=0')
+    response = requests.get('{0}/catalog/data/catalog/digital_objects/?query={{"filter":{{"department":"DigiLab","project":{{"$ne":"private"}},"locations.s3.exists":{{"$eq":true}},"derivatives.jpeg_040_antialias.recipe":{{"$exists":false,"$error":ne}}}}}}&format=json&page_size=0'.format(api_url))
     jobj = response.json()
     results=jobj.get('results')
     for obj in results:
@@ -165,16 +165,25 @@ def read_source_update_derivative(bags,s3_source="source",s3_destination="deriva
     if type(bags) == 'str':
         bags = [bags]
     for bag in bags:
-        task_id = str(read_source_update_derivative.request.id)
         format_params = _params_as_string(outformat,filter,scale,crop)
-
         path_to_bag = "{0}/{1}/{2}/".format(mount_point,s3_source,bag)
         mmsid =get_mmsid(bag,path_to_bag)
         if mmsid:
             bags_with_mmsids[bag]=OrderedDict()
             bags_with_mmsids[bag]['mmsid']=mmsid
-            path_to_tif_files_of_bag = "{0}/{1}/{2}/data/*.tif".format(mount_point,s3_source,bag)
-            print(path_to_tif_files_of_bag)
+            file_extensions = ["*.tif","*.TIFF","*.TIF","*.tiff"]
+            file_paths=[]
+            path_to_manifest_file = "{0}/{1}/{2}/data/manifest*.txt".format(mount_point,s3_source,bag)
+            path_to_bag = "{0}/{1}/{2}/data/".format(mount_point, s3_source, bag)
+            for file in glob.glob(path_to_manifest_file):
+                if(getIntersection(file)):
+                    logging.error("Conflict in bag - {0} : Ambiguous file names (eg. 001.tif , 001.tiff)")
+                else:
+                    for ext in file_extensions:
+                        file_paths.extend(glob.glob(os.path.join((path_to_bag,ext))))
+            if file_paths is None:
+                continue;
+            print(file_paths)
             outdir = "{0}/{1}/{2}/{3}".format(mount_point,s3_destination,bag,format_params)
             if os.path.exists("{0}/derivative/{1}/{2}".format(mount_point, bag, format_params)) and force_overwrite:
                 rmtree(outdir)
@@ -182,14 +191,41 @@ def read_source_update_derivative(bags,s3_source="source",s3_destination="deriva
                 raise derivative_generation_error("derivative already exists and force_overwrite is set to false")
             if not os.path.exists("{0}/derivative/{1}/{2}".format(mount_point, bag, format_params)):
                 os.makedirs(outdir)
-            for file in glob.glob(path_to_tif_files_of_bag):
+            for file in glob.glob(file_paths):
                 print(file)
                 outpath = '{0}/{1}/{2}/{3}/{4}.{5}'.format(mount_point,"derivative",bag,format_params,file.split('/')[-1].split('.')[0].lower(),_formatextension(outformat))
                 processimage(inpath=file,outpath=outpath,outformat=outformat,filter=filter,scale=scale,crop=crop)
         else:
             update_catalog(bag,format_params,mmsid)
-    return {"local_derivatives": "{0}/oulib_tasks/{1}".format(base_url, task_id), "s3_destination": s3_destination,
-            "task_id": task_id,"bags":bags_with_mmsids,"format_params":format_params}
+    return {"s3_destination": s3_destination,
+            "bags":bags_with_mmsids,"format_params":format_params}
+
+
+
+def getIntersection(file):
+    """
+    This method is used for knowing if their are any name conflicts inside the manifest file of each bag.
+    name conflict examples - [ 001.tif , 001.tiff] or [001.TIF , 001.tiff] or [001.TIFF , 001.tif] or [001.TIF , 001.TIFF]
+    :param file: manifest file of each bag.
+    :return: boolean true or false.
+    """
+    ListOfFileNames = []
+    with open(file, "r") as f:
+        lines = f.readlines();
+        for line in lines:
+            line = line.strip();
+            ListOfFileNames.append(line.split("/")[-1])
+    print(len(ListOfFileNames))
+    listOfFileExtensions = [".tif",".tiff"]
+    listOfTif = [name.split(".")[0] for name in ListOfFileNames if name.endswith(".tif") or name.endswith(".TIF")]
+    listOfTiff = [name.split(".")[0] for name in ListOfFileNames if name.endswith(".tiff") or name.endswith(".TIFF")]
+    print(listOfTif)
+    print(listOfTiff)
+    intersectionList = list(set(listOfTiff) & set(listOfTif))
+    if(len(intersectionList) == 0):
+        return False;
+    else:
+        return True;
 
 @task
 def process_recipe(derivative_args):
